+++ src/Polly.Net40/Polly.Net40.csproj
@@ -108,7 +108,6 @@
     <Compile Include="..\Polly.Net35\Utilities\TimedLock.cs">
       <Link>Utilities\TimedLock.cs</Link>
     </Compile>
-    <Compile Include="HandledPolicy.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
